Revert "feat(theme): refactor, theme switch, structure"
