+++ package.json
@@ -15,7 +15,6 @@
   "dependencies": {
     "@material-ui/core": "^4.10.0",
     "@material-ui/icons": "^4.9.1",
-    "dayjs": "^1.9.1",
     "gatsby": "^2.21.37",
     "gatsby-awesome-pagination": "^0.3.6",
     "gatsby-image": "^2.4.5",
@@ -26,7 +25,6 @@
     "gatsby-transformer-json": "^2.4.3",
     "gatsby-transformer-sharp": "^2.5.3",
     "react": "^16.12.0",
-    "react-awesome-button": "^6.5.1",
     "react-dom": "^16.12.0",
     "react-helmet": "^6.0.0"
   },
