--- src/components/Footer/Footer.styles.js
@@ -1,8 +1,0 @@
-import { makeStyles } from '@material-ui/core'
-
-export default makeStyles((theme) => ({
-  footer: {
-    backgroundColor: theme.palette.background.default,
-    padding: theme.spacing(3),
-  },
-}))
