--- src/components/Profile/Profile.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-function Profile() {
-  return <p>Profile</p>
-}
-
-export default Profile
