--- src/components/Profile/Profile.styles.js
@@ -1,3 +1,0 @@
-import { makeStyles } from '@material-ui/core'
-
-export default makeStyles(() => ({}))
