--- src/data/contributors/adnaanbheda.json
@@ -1,7 +1,0 @@
-{
-  "name": "Adnaan Bheda",
-  "desc": "Full-Stack Generalist",
-  "github": "https://github.com/adnaanbheda",
-  "twitter": "https://twitter.com/BhedaAdnaan",
-  "linkedin": "https://www.linkedin.com/in/adnaanb"
-}
